add passcode entry and validation
Got the inputs to auto select and then move to the next input upon entering something.
code validation is in
next is making things happen if you enter the correct or incorrect code, and generating new codes.
And making it erase entered text and go back to the first digit if you change channels (I think? what is desirable here?)