Adding Endless jumper minigame
Endless jumper minigame created. Needs to be added to main scene
Use space to jump, last 60 seconds to get the passcode. Art is pretty rough but idk it might be the best we have time for. Still needs to be hooked into new passcode system and get the new world UI